Update Score v2.2.0 files

+++ score-http/score-http/src/main/java/org/oagi/score/repo/component/ascc/AsccWriteRepository.java
@@ -4,16 +4,17 @@ import org.jooq.DSLContext;
 import org.jooq.UpdateSetFirstStep;
 import org.jooq.UpdateSetMoreStep;
 import org.jooq.types.ULong;
-import org.oagi.score.service.common.data.AppUser;
-import org.oagi.score.service.log.model.LogAction;
 import org.oagi.score.gateway.http.api.cc_management.data.CcASCCPType;
-import org.oagi.score.service.common.data.CcState;
 import org.oagi.score.gateway.http.configuration.security.SessionService;
 import org.oagi.score.gateway.http.helper.ScoreGuid;
-import org.oagi.score.service.log.LogRepository;
 import org.oagi.score.repo.api.ScoreRepositoryFactory;
 import org.oagi.score.repo.api.impl.jooq.entity.tables.records.*;
+import org.oagi.score.service.common.data.AppUser;
+import org.oagi.score.service.common.data.CcState;
+import org.oagi.score.service.corecomponent.seqkey.MoveTo;
 import org.oagi.score.service.corecomponent.seqkey.SeqKeyHandler;
+import org.oagi.score.service.log.LogRepository;
+import org.oagi.score.service.log.model.LogAction;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,11 +22,13 @@ import org.springframework.security.core.AuthenticatedPrincipal;
 import org.springframework.stereotype.Repository;
 
 import java.time.LocalDateTime;
+import java.util.*;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 
 import static org.apache.commons.lang3.StringUtils.compare;
 import static org.jooq.impl.DSL.and;
-import static org.oagi.score.repo.api.impl.jooq.entity.Tables.ASCCP;
-import static org.oagi.score.repo.api.impl.jooq.entity.Tables.ASCCP_MANIFEST;
+import static org.oagi.score.repo.api.impl.jooq.entity.Tables.*;
 import static org.oagi.score.repo.api.impl.jooq.entity.tables.Acc.ACC;
 import static org.oagi.score.repo.api.impl.jooq.entity.tables.AccManifest.ACC_MANIFEST;
 import static org.oagi.score.repo.api.impl.jooq.entity.tables.Ascc.ASCC;
@@ -302,10 +305,28 @@ public class AsccWriteRepository {
             throw new IllegalArgumentException("It only allows to modify the core component by the owner.");
         }
 
+        int usedBieCount = dslContext.selectCount().from(ASBIE)
+                .where(ASBIE.BASED_ASCC_MANIFEST_ID.eq(asccManifestRecord.getAsccManifestId())).fetchOne(0, int.class);
+
+        if (usedBieCount > 0) {
+            throw new IllegalArgumentException("This association used in " + usedBieCount + " BIE(s). Can not be deleted.");
+        }
+
         // delete from Tables
         seqKeyHandler(request.getUser(), asccManifestRecord).deleteCurrent();
+        dslContext.update(ASCC_MANIFEST).setNull(ASCC_MANIFEST.NEXT_ASCC_MANIFEST_ID)
+                .where(ASCC_MANIFEST.NEXT_ASCC_MANIFEST_ID.eq(asccManifestRecord.getAsccManifestId())).execute();
+        dslContext.update(ASCC_MANIFEST).setNull(ASCC_MANIFEST.PREV_ASCC_MANIFEST_ID)
+                .where(ASCC_MANIFEST.PREV_ASCC_MANIFEST_ID.eq(asccManifestRecord.getAsccManifestId())).execute();
         asccManifestRecord.delete();
-        asccRecord.delete();
+        if (dslContext.selectCount().from(ASCC_MANIFEST)
+                .where(ASCC_MANIFEST.ASCC_ID.eq(asccManifestRecord.getAsccId())).fetchOne(0, int.class) == 0) {
+            dslContext.update(ASCC).setNull(ASCC.NEXT_ASCC_ID)
+                    .where(ASCC.NEXT_ASCC_ID.eq(asccRecord.getAsccId())).execute();
+            dslContext.update(ASCC).setNull(ASCC.PREV_ASCC_ID)
+                    .where(ASCC.PREV_ASCC_ID.eq(asccRecord.getAsccId())).execute();
+            asccRecord.delete();
+        }
 
         upsertLogIntoAccAndAssociations(
                 accRecord, accManifestRecord,
@@ -324,5 +345,147 @@ public class AsccWriteRepository {
                 (asccManifestRecord.getSeqKeyId() != null) ? asccManifestRecord.getSeqKeyId().toBigInteger() : null,
                 asccManifestRecord.getAsccManifestId().toBigInteger());
         return seqKeyHandler;
+    }
+
+    public RefactorAsccRepositoryResponse refactor(RefactorAsccRepositoryRequest request) {
+        AppUser user = sessionService.getAppUser(request.getUser());
+        ULong userId = ULong.valueOf(user.getAppUserId());
+        LocalDateTime timestamp = request.getLocalDateTime();
+
+        AsccManifestRecord asccManifestRecord = dslContext.selectFrom(ASCC_MANIFEST)
+                .where(ASCC_MANIFEST.ASCC_MANIFEST_ID.eq(
+                        ULong.valueOf(request.getAsccManifestId())
+                ))
+                .fetchOne();
+
+        AsccRecord asccRecord = dslContext.selectFrom(ASCC)
+                .where(ASCC.ASCC_ID.eq(asccManifestRecord.getAsccId()))
+                .fetchOne();
+
+        AccManifestRecord prevAccManifestRecord = dslContext.selectFrom(ACC_MANIFEST)
+                .where(ACC_MANIFEST.ACC_MANIFEST_ID.eq(asccManifestRecord.getFromAccManifestId()))
+                .fetchOne();
+
+        AccRecord prevAccRecord = dslContext.selectFrom(ACC)
+                .where(ACC.ACC_ID.eq(prevAccManifestRecord.getAccId()))
+                .fetchOne();
+
+        AccManifestRecord accManifestRecord = dslContext.selectFrom(ACC_MANIFEST)
+                .where(ACC_MANIFEST.ACC_MANIFEST_ID.eq(ULong.valueOf(request.getAccManifestId())))
+                .fetchOne();
+
+        AccRecord accRecord = dslContext.selectFrom(ACC)
+                .where(ACC.ACC_ID.eq(accManifestRecord.getAccId()))
+                .fetchOne();
+
+        String asccpDen = dslContext.select(ASCCP.DEN)
+                .from(ASCCP_MANIFEST)
+                .join(ASCCP).on(ASCCP_MANIFEST.ASCCP_ID.eq(ASCCP.ASCCP_ID))
+                .where(ASCCP_MANIFEST.ASCCP_MANIFEST_ID.eq(asccManifestRecord.getToAsccpManifestId()))
+                .fetchOneInto(String.class);
+
+        if (!CcState.WIP.equals(CcState.valueOf(accRecord.getState()))) {
+            throw new IllegalArgumentException("Only the core component in 'WIP' state can be refactored.");
+        }
+
+        if (!accRecord.getOwnerUserId().equals(userId)) {
+            throw new IllegalArgumentException("It only allows to modify the core component by the owner.");
+        }
+
+
+        int usedBieCount = dslContext.selectCount().from(ASBIE)
+                .where(ASBIE.BASED_ASCC_MANIFEST_ID.eq(asccManifestRecord.getAsccManifestId())).fetchOne(0, int.class);
+
+        if (usedBieCount > 0) {
+            throw new IllegalArgumentException("This association used in " + usedBieCount + " BIE(s). Can not be refactored.");
+        }
+
+        this.checkDuplicate(asccManifestRecord, accManifestRecord.getAccManifestId());
+
+        // delete from Tables
+        seqKeyHandler(request.getUser(), asccManifestRecord).deleteCurrent();
+        asccManifestRecord.setFromAccManifestId(ULong.valueOf(request.getAccManifestId()));
+        asccManifestRecord.update(ASCC_MANIFEST.FROM_ACC_MANIFEST_ID);
+
+        seqKeyHandler(request.getUser(), asccManifestRecord).moveTo(MoveTo.LAST);
+
+        asccRecord.setLastUpdatedBy(userId);
+        asccRecord.setLastUpdateTimestamp(timestamp);
+        asccRecord.setFromAccId(accRecord.getAccId());
+        asccRecord.setDen(accRecord.getObjectClassTerm() + ". " + asccpDen);
+        asccRecord.update(ASCC.LAST_UPDATED_BY, ASCC.LAST_UPDATE_TIMESTAMP,
+                ASCC.FROM_ACC_ID, ASCC.DEN);
+
+        upsertLogIntoAccAndAssociations(
+                accRecord, accManifestRecord,
+                accManifestRecord.getReleaseId(),
+                userId, timestamp
+        );
+
+        upsertLogIntoAccAndAssociations(
+                prevAccRecord, prevAccManifestRecord,
+                accManifestRecord.getReleaseId(),
+                userId, timestamp
+        );
+
+        return new RefactorAsccRepositoryResponse(asccManifestRecord.getAsccManifestId().toBigInteger());
+    }
+
+    private void checkDuplicate(AsccManifestRecord asccManifestRecord, ULong targetAccManifestId) {
+        List<AccManifestRecord> accList = dslContext.selectFrom(ACC_MANIFEST)
+                .where(ACC_MANIFEST.RELEASE_ID.eq(asccManifestRecord.getReleaseId())).fetch();
+
+        Map<ULong, AccManifestRecord> accMap = accList.stream().collect(Collectors.toMap(AccManifestRecord::getAccManifestId, Function.identity()));
+        Map<ULong, List<AccManifestRecord>> baseAccMap = accList.stream().filter(e -> e.getBasedAccManifestId() != null)
+                .collect(Collectors.groupingBy(AccManifestRecord::getBasedAccManifestId));
+
+        List<AsccManifestRecord> asccList = dslContext.selectFrom(ASCC_MANIFEST)
+                .where(ASCC_MANIFEST.RELEASE_ID.eq(asccManifestRecord.getReleaseId())).fetch();
+
+        Map<ULong, List<AsccManifestRecord>> fromAccAsccMap = asccList.stream()
+                .collect(Collectors.groupingBy(AsccManifestRecord::getFromAccManifestId));
+
+        List<ULong> accManifestList = new ArrayList<>();
+
+        AccManifestRecord targetAccManifestRecord = accMap.get(targetAccManifestId);
+
+        accManifestList.add(targetAccManifestId);
+
+        while (targetAccManifestRecord.getBasedAccManifestId() != null) {
+            accManifestList.add(targetAccManifestRecord.getBasedAccManifestId());
+            targetAccManifestRecord = accMap.get(targetAccManifestRecord.getBasedAccManifestId());
+        }
+
+        Set<ULong> result = new HashSet<>();
+
+        for (ULong cur : accManifestList) {
+            result.addAll(getBaseAccManifestId(cur, baseAccMap));
+        }
+
+        Set<AsccManifestRecord> asccResult = new HashSet<>();
+
+        for (ULong acc : result) {
+            asccResult.addAll(
+                    fromAccAsccMap.getOrDefault(acc, Collections.emptyList())
+                            .stream()
+                            .filter(ascc -> ascc.getToAsccpManifestId().equals(asccManifestRecord.getToAsccpManifestId())
+                                    && !ascc.getAsccManifestId().equals(asccManifestRecord.getAsccManifestId()))
+                            .collect(Collectors.toList()));
+        }
+
+        if (asccResult.size() > 0) {
+            throw new IllegalArgumentException("Unable to refactor. " + asccResult.size() + " duplicate association(s) exist in the derived ACC");
+        }
+    }
+
+    private List<ULong> getBaseAccManifestId(ULong accManifestId, Map<ULong, List<AccManifestRecord>> baseAccMap) {
+        List<ULong> result = new ArrayList<>();
+        result.add(accManifestId);
+        if (baseAccMap.containsKey(accManifestId)) {
+            baseAccMap.get(accManifestId).forEach(e -> {
+                result.addAll(getBaseAccManifestId(e.getAccManifestId(), baseAccMap));
+            });
+        }
+        return result;
     }
 }
